Properly classify progression item instances and rupees in log file
So that only the earliest minimum required rupees are listed as
progression, and the earliest instance of a required item class is
always progression. This technically doesn't cover cases where two items
of different classes are both progression but either is skippable, but
that case is rare enough that I don't think covering it is worth the
effort and added complexity.