*Do* use fanfare + animation + text box for non-falling SKs
Previous behavior was to use text, but no fanfare or animation - a
terrible in-between. I tried using no text, but it's confusing if it
happens when you're standing on the spawn point for a boss HC and you
get no text or animation, and can't hear the pickup sound over the boss
dying.